chore: separates songname and settings
